+++ src/modem/freq.rs
@@ -16,12 +16,12 @@ impl Freq {
     }
 
     // Get radians per second.
-    pub fn ang_freq(&self) -> f64 {
-        2.0 * std::f64::consts::PI * self.hz as f64
+    pub fn ang_freq(&self) -> f32 {
+        2.0 * std::f32::consts::PI * self.hz as f32
     }
 
     // Get radians per sample given the samples per second.
-    pub fn sample_freq(&self) -> f64 {
-        self.ang_freq() / self.sr as f64
+    pub fn sample_freq(&self) -> f32 {
+        self.ang_freq() / self.sr as f32
     }
 }
