+++ src/modem/util.rs
@@ -1,7 +1,7 @@
 use std;
 
-pub fn mod_trig(x: f64) -> f64 {
-    const TWO_PI: f64 = std::f64::consts::PI * 2.0;
+pub fn mod_trig(x: f32) -> f32 {
+    const TWO_PI: f32 = std::f32::consts::PI * 2.0;
 
     x - TWO_PI * (x / TWO_PI).floor()
 }
